Modify all folders
